+++ linero/table/crossed.py
@@ -16,7 +16,7 @@ Strategy:
     1. Add buckets with horizontal lines
     2. Iter thru vertical lines and add boundings in every hitted bucket
     3. Select connected buckets
-    4. Connect small table framents which are next and close to each other
+    4. Connect small table fragments which are next and close to each other
 """
 
 import operator